+++ CHANGELOG.md
@@ -3,13 +3,17 @@
 
 **Features:**
 - Added #1171: Show start writing message when no posts.
+- Added #1173: Discord link to dashboard.
 
 **Bug Fixes:**
 - Fixed #1155: Fix top margin on main wrapper.
 - Fixed #1153: Fix dark mode on reader and post/page edition.
+- Fixed #1129: Sidebar link is still active when in account state.
+- Fixed #1180: Hide sidebar when navigating through 'site-new' state.
 
 **Improvement:**
 - Closes #1158: Google AdService Conversion.
+- CLoses #1162: Move post status tabs next to search bar to save space.
 
 
 2019.4.1